# ddl: Speed up adding index phase