--- a/ddl/index.go
+++ b/ddl/index.go
@@ -418,22 +418,21 @@ type indexBatchOpInfo struct {
 }
 
 // How to add index in reorganization state?
-// Concurrently process defaultSmallBatches tasks. Each task deals with a handle interval of the index record.
-// The handle interval is defaultSmallBatchCnt.
-// Although the length of each handle interval is controllable, but the range of the handle value can't be expected,
-// so it's necessary to obtain the handle range serially. Real concurrent processing needs to perform
-// after the handle interval has been acquired.
+// Concurrently process defaultSmallBatches tasks. Each task deals with a handle range of the index record.
+// The handle range size is defaultSmallBatchCnt.
+// Because each handle range depends on the previous one, it's necessary to obtain the handle range serially.
+// Real concurrent processing needs to perform after the handle range has been acquired.
 // The operation flow of the each batch of data is as follows:
 //  1. Open a goroutine. Traverse the snapshot to obtain the handle range, while access to the corresponding row key and
 // raw index value. Then notify to start the next batch.
-//  2. Decoding this batch of raw index value, gets the corresponding index value.
-//  3. Deal with this index records one by one. If the index record is existed, skip to the next row.
+//  2. Decoding this batch of raw index value gets the corresponding index value.
+//  3. Deal with this index records one by one. If the index record exists, skip to the next row.
 // If the index doesn't exist, create the index ande then continue to handle the next row.
-//  4. When the handle of range is completed, returns the corresponding batch result.
+//  4. When the handle of a range is completed, returns the corresponding batch result.
 // The above operations are completed in a transaction.
-// When concurrent tasks are processed, the batch result returned by each batch is sorted by handle. Then traverse the
+// When concurrent tasks are processed, the batch result returned by each batch is sorted by the handle. Then traverse the
 // batch results, gets the total number of row in the concurrent task and update the processed handle value. If
-// you encounter an error message, exit traversal.
+// we encounter an error message, exit traversal.
 // Finally, update the concurrent processing of the total number of rows, and store the completed handle value.
 func (d *ddl) addTableIndex(t table.Table, indexInfo *model.IndexInfo, reorgInfo *reorgInfo, job *model.Job) error {
 	cols := t.Cols()