--- a/ddl/index.go
+++ b/ddl/index.go
@@ -14,6 +14,8 @@
 package ddl
 
 import (
+	"sort"
+	"sync"
 	"time"
 
 	"github.com/juju/errors"
@@ -324,43 +326,99 @@ func (d *ddl) onDropIndex(t *meta.Meta, job *model.Job) error {
 	return errors.Trace(err)
 }
 
-func (d *ddl) fetchRowColVals(txn kv.Transaction, t table.Table, batchOpInfo *indexBatchOpInfo, seekHandle int64) error {
-	cols := t.Cols()
-	idxInfo := batchOpInfo.tblIndex.Meta()
+func isFinish(limit, input int64) bool {
+	if limit == 0 || input < limit {
+		return false
+	}
+	return true
+}
 
-	err := d.iterateSnapshotRows(t, txn.StartTS(), seekHandle,
+func (d *ddl) fetchRowColVals(txn kv.Transaction, t table.Table, batchOpInfo *indexBatchOpInfo, handleInfo *handleInfo) (
+	[]*indexRecord, *batchRet) {
+	batchCnt := defaultSmallBatchCnt
+	rawRecords := make([][]byte, 0, batchCnt)
+	idxRecords := make([]*indexRecord, 0, batchCnt)
+	ret := new(batchRet)
+	err := d.iterateSnapshotRows(t, txn.StartTS(), handleInfo.startHandle,
 		func(h int64, rowKey kv.Key, rawRecord []byte) (bool, error) {
-			rowMap, err := tablecodec.DecodeRow(rawRecord, batchOpInfo.colMap)
-			if err != nil {
-				return false, errors.Trace(err)
-			}
-			idxVal := make([]types.Datum, 0, len(idxInfo.Columns))
-			for _, v := range idxInfo.Columns {
-				col := cols[v.Offset]
-				idxVal = append(idxVal, rowMap[col.ID])
-			}
-
-			indexRecord := &indexRecord{handle: h, key: rowKey, vals: idxVal}
-			batchOpInfo.idxRecords = append(batchOpInfo.idxRecords, indexRecord)
-			if len(batchOpInfo.idxRecords) == defaultSmallBatchCnt {
+			rawRecords = append(rawRecords, rawRecord)
+			indexRecord := &indexRecord{handle: h, key: rowKey}
+			idxRecords = append(idxRecords, indexRecord)
+			if len(idxRecords) == batchCnt || isFinish(handleInfo.endHandle, h) {
 				return false, nil
 			}
 			return true, nil
 		})
 	if err != nil {
-		return errors.Trace(err)
-	} else if len(batchOpInfo.idxRecords) == 0 {
-		return nil
+		ret.err = errors.Trace(err)
+		return nil, ret
 	}
 
-	count := len(batchOpInfo.idxRecords)
-	batchOpInfo.addedCount += int64(count)
-	batchOpInfo.handle = batchOpInfo.idxRecords[count-1].handle
-	return nil
+	ret.count = len(idxRecords)
+	if ret.count > 0 {
+		ret.doneHandle = idxRecords[ret.count-1].handle
+	}
+	// Be sure to do this operation only once.
+	handleInfo.once.Do(func() {
+		// Notice to start the next batch operation.
+		batchOpInfo.nextCh <- ret.doneHandle
+		// Record the last handle.
+		// Ensure that the handle scope of the batch doesn't change,
+		// even if the transaction retries it can't effect the other batches.
+		handleInfo.endHandle = ret.doneHandle
+	})
+	if ret.count == 0 {
+		return nil, ret
+	}
+
+	cols := t.Cols()
+	idxInfo := batchOpInfo.tblIndex.Meta()
+	for i, idxRecord := range idxRecords {
+		rowMap, err := tablecodec.DecodeRow(rawRecords[i], batchOpInfo.colMap)
+		if err != nil {
+			ret.err = errors.Trace(err)
+			return nil, ret
+		}
+		idxVal := make([]types.Datum, 0, len(idxInfo.Columns))
+		for _, v := range idxInfo.Columns {
+			col := cols[v.Offset]
+			idxVal = append(idxVal, rowMap[col.ID])
+		}
+		idxRecord.vals = idxVal
+	}
+	return idxRecords, ret
 }
 
-const defaultBatchCnt = 1024
-const defaultSmallBatchCnt = 128
+const (
+	defaultBatchCnt      = 1024
+	defaultSmallBatchCnt = 128
+	defaultSmallBatches  = 16
+)
+
+type batchRet struct {
+	count      int
+	doneHandle int64 // This is the last reorg handle that has been processed.
+	err        error
+}
+
+type batchRetSlice []*batchRet
+
+func (b batchRetSlice) Len() int           { return len(b) }
+func (b batchRetSlice) Less(i, j int) bool { return b[i].doneHandle < b[j].doneHandle }
+func (b batchRetSlice) Swap(i, j int)      { b[i], b[j] = b[j], b[i] }
+
+type indexRecord struct {
+	handle int64
+	key    []byte
+	vals   []types.Datum
+}
+
+type indexBatchOpInfo struct {
+	tblIndex   table.Index
+	colMap     map[int64]*types.FieldType
+	batchRetCh chan *batchRet
+	nextCh     chan int64
+}
 
 // How to add index in reorganization state?
 //  1. Generate a snapshot with special version.
@@ -375,46 +433,173 @@ func (d *ddl) addTableIndex(t table.Table, indexInfo *model.IndexInfo, reorgInfo
 		col := cols[v.Offset]
 		colMap[col.ID] = &col.FieldType
 	}
-	batchCnt := defaultSmallBatchCnt
+	batches := defaultSmallBatches
 	batchOpInfo := &indexBatchOpInfo{
 		tblIndex:   tables.NewIndex(t.Meta(), indexInfo),
-		addedCount: job.GetRowCount(),
 		colMap:     colMap,
-		handle:     reorgInfo.Handle,
-		idxRecords: make([]*indexRecord, 0, batchCnt),
+		nextCh:     make(chan int64, 1),
+		batchRetCh: make(chan *batchRet, batches),
 	}
 
+	addedCount := job.GetRowCount()
 	seekHandle := reorgInfo.Handle
+	wg := sync.WaitGroup{}
 	for {
 		startTime := time.Now()
-		err := kv.RunInNewTxn(d.store, true, func(txn kv.Transaction) error {
-			err1 := d.isReorgRunnable(txn, ddlJobFlag)
-			if err1 != nil {
-				return errors.Trace(err1)
+		for i := 0; i < batches; i++ {
+			go d.backfillIndex(t, batchOpInfo, seekHandle, &wg)
+			handle := <-batchOpInfo.nextCh
+			// There is no data to seek.
+			if handle == 0 {
+				break
 			}
-			batchOpInfo.idxRecords = batchOpInfo.idxRecords[:0]
-			err1 = d.backfillIndexInTxn(t, txn, batchOpInfo, seekHandle)
+			seekHandle = handle + 1
+		}
+		wg.Wait()
+
+		retCnt := len(batchOpInfo.batchRetCh)
+		batchAddedCount, doneHandle, err := getCountAndHandle(batchOpInfo)
+		// Update the reorg handle that has been processed.
+		if batchAddedCount != 0 {
+			err1 := kv.RunInNewTxn(d.store, true, func(txn kv.Transaction) error {
+				return errors.Trace(reorgInfo.UpdateHandle(txn, doneHandle+1))
+			})
 			if err1 != nil {
-				return errors.Trace(err1)
+				if err == nil {
+					err = err1
+				} else {
+					log.Warnf("[ddl] add index failed when update handle %d, err %v", doneHandle, err)
+				}
 			}
-			// Update the reorg handle that has been processed.
-			return errors.Trace(reorgInfo.UpdateHandle(txn, batchOpInfo.handle))
-		})
+		}
+
+		addedCount += int64(batchAddedCount)
 		sub := time.Since(startTime).Seconds()
 		if err != nil {
-			log.Warnf("[ddl] added index for %v rows failed, take time %v", batchOpInfo.addedCount, sub)
+			log.Warnf("[ddl] total added index for %d rows, this batch add index for %d failed, take time %v",
+				addedCount, batchAddedCount, sub)
 			return errors.Trace(err)
 		}
-
-		job.SetRowCount(batchOpInfo.addedCount)
+		job.SetRowCount(addedCount)
 		batchHandleDataHistogram.WithLabelValues(batchAddIdx).Observe(sub)
-		log.Infof("[ddl] added index for %v rows, take time %v", batchOpInfo.addedCount, sub)
+		log.Infof("[ddl] total added index for %d rows, this batch added index for %d rows, take time %v",
+			addedCount, batchAddedCount, sub)
 
-		if len(batchOpInfo.idxRecords) < batchCnt {
+		if retCnt < batches {
 			return nil
 		}
-		seekHandle = batchOpInfo.handle + 1
 	}
+}
+
+type handleInfo struct {
+	startHandle int64
+	endHandle   int64
+	once        sync.Once
+}
+
+func getCountAndHandle(batchOpInfo *indexBatchOpInfo) (int64, int64, error) {
+	l := len(batchOpInfo.batchRetCh)
+	batchRets := make([]*batchRet, 0, l)
+	for i := 0; i < l; i++ {
+		batchRet := <-batchOpInfo.batchRetCh
+		batchRets = append(batchRets, batchRet)
+	}
+	sort.Sort(batchRetSlice(batchRets))
+
+	batchAddedCount, currHandle := int64(0), int64(0)
+	var err error
+	for _, ret := range batchRets {
+		if ret.err != nil {
+			err = ret.err
+			break
+		}
+		batchAddedCount += int64(ret.count)
+		currHandle = ret.doneHandle
+	}
+	return batchAddedCount, currHandle, errors.Trace(err)
+}
+
+func (d *ddl) backfillIndex(t table.Table, batchOpInfo *indexBatchOpInfo, seekHandle int64, wg *sync.WaitGroup) {
+	wg.Add(1)
+	defer wg.Done()
+
+	ret := new(batchRet)
+	handleInfo := &handleInfo{startHandle: seekHandle}
+	err := kv.RunInNewTxn(d.store, true, func(txn kv.Transaction) error {
+		err1 := d.isReorgRunnable(txn, ddlJobFlag)
+		if err1 != nil {
+			return errors.Trace(err1)
+		}
+		ret = d.backfillIndexInTxn(t, txn, batchOpInfo, handleInfo)
+		if ret.err != nil {
+			return errors.Trace(ret.err)
+		}
+		return nil
+	})
+	if err != nil {
+		ret.err = errors.Trace(err)
+	}
+
+	// It's failed to fetch row keys.
+	if ret.count == 0 && ret.err != nil {
+		batchOpInfo.nextCh <- seekHandle
+	}
+
+	batchOpInfo.batchRetCh <- ret
+}
+
+// backfillIndexInTxn deals with a part of backfilling index data in a Transaction.
+// This part of the index data rows is defaultSmallBatchCnt.
+func (d *ddl) backfillIndexInTxn(t table.Table, txn kv.Transaction, batchOpInfo *indexBatchOpInfo,
+	handleInfo *handleInfo) *batchRet {
+	idxRecords, batchRet := d.fetchRowColVals(txn, t, batchOpInfo, handleInfo)
+	if batchRet.err != nil {
+		batchRet.err = errors.Trace(batchRet.err)
+		return batchRet
+	}
+
+	for _, idxRecord := range idxRecords {
+		log.Debug("[ddl] backfill index...", idxRecord.handle)
+		err := txn.LockKeys(idxRecord.key)
+		if err != nil {
+			batchRet.err = errors.Trace(err)
+			return batchRet
+		}
+
+		// Create the index.
+		handle, err := batchOpInfo.tblIndex.Create(txn, idxRecord.vals, idxRecord.handle)
+		if err != nil {
+			if terror.ErrorEqual(err, kv.ErrKeyExists) && idxRecord.handle == handle {
+				// Index already exists, skip it.
+				continue
+			}
+			batchRet.err = errors.Trace(err)
+			return batchRet
+		}
+	}
+	return batchRet
+}
+
+func (d *ddl) dropTableIndex(indexInfo *model.IndexInfo, job *model.Job) error {
+	startKey := tablecodec.EncodeTableIndexPrefix(job.TableID, indexInfo.ID)
+	// It's asynchronous so it doesn't need to consider if it completes.
+	deleteAll := -1
+	_, _, err := d.delKeysWithStartKey(startKey, startKey, ddlJobFlag, job, deleteAll)
+	return errors.Trace(err)
+}
+
+func findIndexByName(idxName string, indices []*model.IndexInfo) *model.IndexInfo {
+	for _, idx := range indices {
+		if idx.Name.L == idxName {
+			return idx
+		}
+	}
+	return nil
+}
+
+func allocateIndexID(tblInfo *model.TableInfo) int64 {
+	tblInfo.MaxIndexID++
+	return tblInfo.MaxIndexID
 }
 
 // recordIterFunc is used for low-level record iteration.
@@ -460,68 +645,4 @@ func (d *ddl) iterateSnapshotRows(t table.Table, version uint64, seekHandle int6
 	}
 
 	return nil
-}
-
-type indexBatchOpInfo struct {
-	tblIndex   table.Index
-	addedCount int64
-	handle     int64 // This is the last reorg handle that has been processed.
-	colMap     map[int64]*types.FieldType
-	idxRecords []*indexRecord
-}
-
-type indexRecord struct {
-	handle int64
-	key    []byte
-	vals   []types.Datum
-}
-
-// backfillIndexInTxn deals with a part of backfilling index data in a Transaction.
-// This part of the index data rows is defaultSmallBatchCnt.
-func (d *ddl) backfillIndexInTxn(t table.Table, txn kv.Transaction, batchOpInfo *indexBatchOpInfo, seekHandle int64) error {
-	err := d.fetchRowColVals(txn, t, batchOpInfo, seekHandle)
-	if err != nil {
-		return errors.Trace(err)
-	}
-
-	for _, idxRecord := range batchOpInfo.idxRecords {
-		log.Debug("[ddl] backfill index...", idxRecord.handle)
-		err = txn.LockKeys(idxRecord.key)
-		if err != nil {
-			return errors.Trace(err)
-		}
-
-		// Create the index.
-		handle, err := batchOpInfo.tblIndex.Create(txn, idxRecord.vals, idxRecord.handle)
-		if err != nil {
-			if terror.ErrorEqual(err, kv.ErrKeyExists) && idxRecord.handle == handle {
-				// Index already exists, skip it.
-				continue
-			}
-			return errors.Trace(err)
-		}
-	}
-	return nil
-}
-
-func (d *ddl) dropTableIndex(indexInfo *model.IndexInfo, job *model.Job) error {
-	startKey := tablecodec.EncodeTableIndexPrefix(job.TableID, indexInfo.ID)
-	// It's asynchronous so it doesn't need to consider if it completes.
-	deleteAll := -1
-	_, _, err := d.delKeysWithStartKey(startKey, startKey, ddlJobFlag, job, deleteAll)
-	return errors.Trace(err)
-}
-
-func findIndexByName(idxName string, indices []*model.IndexInfo) *model.IndexInfo {
-	for _, idx := range indices {
-		if idx.Name.L == idxName {
-			return idx
-		}
-	}
-	return nil
-}
-
-func allocateIndexID(tblInfo *model.TableInfo) int64 {
-	tblInfo.MaxIndexID++
-	return tblInfo.MaxIndexID
 }