Handle issue #2257
Remove the batch get operation when adding index.

Using 3 TiKV, 1 PD, 1 TiDB , the servers of TiKV and TiDB are not on a computer, two TiKV on a computer.
The lease of TiDB is 1s.
The number of records in this table is 3531200(3.4 M), and the table's structure is as follows:
+------------------+------------------+------+------+---------+-------+
| Field | Type | Null | Key | Default | Extra |
+------------------+------------------+------+------+---------+-------+
| a | varchar(64) | YES | | | |
| b | bigint(20) | NO | | 0 | |
| c | datetime | YES | | NULL | |
| d | int(11) UNSIGNED | NO | | 0 | |
| e | int(11) UNSIGNED | NO | | 0 | |
| f | int(11) UNSIGNED | NO | | 0 | |
| g | varchar(28) | NO | | | |
| h | varchar(28) | NO | | | |
+------------------+------------------+------+------+---------+-------+

tidb-before> alter table battle_begin add index b (event_id);
Query OK, 0 rows affected (8 min 36.40 sec)
tidb-before>  alter table battle_begin add index g (sum_id);
Query OK, 0 rows affected (7 min 10.01 sec)

tidb-current> alter table battle_begin add index b (event_id);
Query OK, 0 rows affected (3 min 35.75 sec)
tidb-current> alter table battle_begin add index g (sum_id);
Query OK, 0 rows affected (3 min 46.25 sec)